+++ packages/components/sidebar/src/sidebar-content.tsx
@@ -1,8 +1,12 @@
-import { cn, type UIProps } from "@jamsr-ui/utils";
+import { cn, type ComponentPropsWithAs } from "@jamsr-ui/utils";
 import { useSidebarContext } from "./sidebar-context";
 
-export type SidebarContentProps = UIProps<"div">;
-export const SidebarContent = (props: SidebarContentProps) => {
+export type SidebarContentProps<T extends React.ElementType = "div"> =
+  ComponentPropsWithAs<T>;
+
+export const SidebarContent = <T extends React.ElementType = "div">(
+  props: SidebarContentProps<T>,
+) => {
   const { children, className: $className, ...restProps } = props;
   const { styles, classNames } = useSidebarContext();
   const className = styles.content({