+++ packages/components/sidebar/src/sidebar-footer.tsx
@@ -1,8 +1,12 @@
-import { cn, type UIProps } from "@jamsr-ui/utils";
+import { cn, type ComponentPropsWithAs } from "@jamsr-ui/utils";
 import { useSidebarContext } from "./sidebar-context";
 
-export type SidebarFooterProps = UIProps<"div">;
-export const SidebarFooter = (props: SidebarFooterProps) => {
+export type SidebarFooterProps<T extends React.ElementType = "div"> =
+  ComponentPropsWithAs<T>;
+
+export const SidebarFooter = <T extends React.ElementType = "div">(
+  props: SidebarFooterProps<T>,
+) => {
   const { children, className: $className, ...restProps } = props;
   const { styles, classNames } = useSidebarContext();
   const className = styles.footer({