+++ packages/components/sidebar/src/sidebar-group.tsx
@@ -1,8 +1,12 @@
-import { cn, type UIProps } from "@jamsr-ui/utils";
+import { cn, type ComponentPropsWithAs } from "@jamsr-ui/utils";
 import { useSidebarContext } from "./sidebar-context";
 
-export type SidebarGroupProps = UIProps<"div">;
-export const SidebarGroup = (props: SidebarGroupProps) => {
+export type SidebarGroupProps<T extends React.ElementType = "div"> =
+  ComponentPropsWithAs<T>;
+
+export const SidebarGroup = <T extends React.ElementType = "div">(
+  props: SidebarGroupProps<T>,
+) => {
   const { children, className: $className, ...restProps } = props;
   const { styles, classNames } = useSidebarContext();
   const className = styles.group({