+++ packages/components/sidebar/src/sidebar-header.tsx
@@ -1,8 +1,12 @@
-import { cn, type UIProps } from "@jamsr-ui/utils";
+import { cn, type ComponentPropsWithAs } from "@jamsr-ui/utils";
 import { useSidebarContext } from "./sidebar-context";
 
-export type SidebarHeaderProps = UIProps<"div">;
-export const SidebarHeader = (props: SidebarHeaderProps) => {
+export type SidebarHeaderProps<T extends React.ElementType = "div"> =
+  ComponentPropsWithAs<T>;
+
+export const SidebarHeader = <T extends React.ElementType = "div">(
+  props: SidebarHeaderProps<T>,
+) => {
   const { children, className: $className, ...restProps } = props;
   const { styles, classNames } = useSidebarContext();
   const className = styles.header({