+++ packages/components/sidebar/src/sidebar-menu-item-button.tsx
@@ -1,15 +1,20 @@
-import { cn, dataAttr, type UIProps } from "@jamsr-ui/utils";
+import { cn, type ComponentPropsWithAs, dataAttr } from "@jamsr-ui/utils";
 import { useSidebarContext } from "./sidebar-context";
 
-export type SidebarMenuItemButtonProps = UIProps<"button"> & {
-  isDisabled?: boolean;
-};
-export const SidebarMenuItemButton = (props: SidebarMenuItemButtonProps) => {
+export type SidebarMenuItemButtonProps<T extends React.ElementType = "button"> =
+  ComponentPropsWithAs<T> & {
+    isDisabled?: boolean;
+  };
+
+export const SidebarMenuItemButton = <T extends React.ElementType = "button">(
+  props: SidebarMenuItemButtonProps<T>,
+) => {
   const {
     children,
     className: $className,
     isDisabled = false,
     disabled: $disabled,
+    as,
     ...restProps
   } = props;
   const { styles, classNames } = useSidebarContext();
@@ -17,8 +22,9 @@ export const SidebarMenuItemButton = (props: SidebarMenuItemButtonProps) => {
     className: cn(classNames?.menuItemButton, $className),
   });
   const disabled = isDisabled || $disabled;
+  const Component = as ?? "button";
   return (
-    <button
+    <Component
       type="button"
       data-slot="menuItemButton"
       className={className}
@@ -28,6 +34,6 @@ export const SidebarMenuItemButton = (props: SidebarMenuItemButtonProps) => {
       {...restProps}
     >
       {children}
-    </button>
+    </Component>
   );
 };