+++ packages/components/sidebar/src/sidebar-menu-item.tsx
@@ -1,8 +1,12 @@
-import { cn, type UIProps } from "@jamsr-ui/utils";
+import { cn, type ComponentPropsWithAs } from "@jamsr-ui/utils";
 import { useSidebarContext } from "./sidebar-context";
 
-export type SidebarMenuItemProps = UIProps<"li">;
-export const SidebarMenuItem = (props: SidebarMenuItemProps) => {
+export type SidebarMenuItemProps<T extends React.ElementType = "li"> =
+  ComponentPropsWithAs<T>;
+
+export const SidebarMenuItem = <T extends React.ElementType = "li">(
+  props: SidebarMenuItemProps<T>,
+) => {
   const { children, className: $className, ...restProps } = props;
   const { styles, classNames } = useSidebarContext();
   const className = styles.menuItem({