+++ packages/components/sidebar/src/sidebar-menu.tsx
@@ -1,8 +1,12 @@
-import { cn, type UIProps } from "@jamsr-ui/utils";
+import { cn, type ComponentPropsWithAs } from "@jamsr-ui/utils";
 import { useSidebarContext } from "./sidebar-context";
 
-export type SidebarMenuProps = UIProps<"ul">;
-export const SidebarMenu = (props: SidebarMenuProps) => {
+export type SidebarMenuProps<T extends React.ElementType = "ul"> =
+  ComponentPropsWithAs<T>;
+
+export const SidebarMenu = <T extends React.ElementType = "div">(
+  props: SidebarMenuProps<T>,
+) => {
   const { children, className: $className, ...restProps } = props;
   const { styles, classNames } = useSidebarContext();
   const className = styles.menu({