+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ui/storybook
 
+## 0.0.222
+
+### Patch Changes
+
+- @jamsr-ui/react@0.0.220
+
 ## 0.0.221
 
 ### Patch Changes